Remove driver file
