+++ pkg/azurelustre/azurelustre.go
@@ -127,9 +127,18 @@ type LustreSkuValue struct {
 	AvailableZones []string
 }
 
+type CSIDriver struct {
+	Name    string
+	NodeID  string
+	Version string
+	Cap     []*csi.ControllerServiceCapability
+	VC      []*csi.VolumeCapability_AccessMode
+	NSCap   []*csi.NodeServiceCapability
+}
+
 // Driver implements all interfaces of CSI drivers
 type Driver struct {
-	csicommon.CSIDriver
+	CSIDriver
 	csi.UnimplementedControllerServer
 	csi.UnimplementedNodeServer
 	csi.UnimplementedIdentityServer
@@ -323,6 +332,34 @@ func (d *Driver) Run(endpoint string, testBool bool) {
 	// Driver d act as IdentityServer, ControllerServer and NodeServer
 	s.Start(endpoint, d, d, d, testBool)
 	s.Wait()
+}
+
+func (d *Driver) AddControllerServiceCapabilities(cl []csi.ControllerServiceCapability_RPC_Type) {
+	csc := make([]*csi.ControllerServiceCapability, 0, len(cl))
+
+	for _, c := range cl {
+		klog.Infof("Enabling controller service capability: %v", c.String())
+		csc = append(csc, NewControllerServiceCapability(c))
+	}
+	d.Cap = csc
+}
+
+func (d *Driver) AddNodeServiceCapabilities(nl []csi.NodeServiceCapability_RPC_Type) {
+	nsc := make([]*csi.NodeServiceCapability, 0, len(nl))
+	for _, n := range nl {
+		klog.V(2).Infof("Enabling node service capability: %v", n.String())
+		nsc = append(nsc, NewNodeServiceCapability(n))
+	}
+	d.NSCap = nsc
+}
+
+func (d *Driver) AddVolumeCapabilityAccessModes(vc []csi.VolumeCapability_AccessMode_Mode) {
+	vca := make([]*csi.VolumeCapability_AccessMode, 0, len(vc))
+	for _, c := range vc {
+		klog.Infof("Enabling volume access mode: %v", c.String())
+		vca = append(vca, NewVolumeCapabilityAccessMode(c))
+	}
+	d.VC = vca
 }
 
 func IsCorruptedDir(dir string) bool {
