--- pkg/csi-common/driver.go
@@ -1,105 +1,0 @@
-/*
-Copyright 2017 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package csicommon
-
-import (
-	"github.com/container-storage-interface/spec/lib/go/csi"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
-	"k8s.io/klog/v2"
-)
-
-type CSIDriver struct {
-	Name    string
-	NodeID  string
-	Version string
-	Cap     []*csi.ControllerServiceCapability
-	VC      []*csi.VolumeCapability_AccessMode
-	NSCap   []*csi.NodeServiceCapability
-}
-
-// Creates a NewCSIDriver object. Assumes vendor version is equal to driver version &
-// does not support optional driver plugin info manifest field. Refer to CSI spec for more details.
-func NewCSIDriver(name, v, nodeID string) *CSIDriver {
-	if name == "" {
-		klog.Errorf("Driver name missing")
-		return nil
-	}
-
-	if nodeID == "" {
-		klog.Errorf("NodeID missing")
-		return nil
-	}
-	// TODO version format and validation
-	if len(v) == 0 {
-		klog.Errorf("Version argument missing, now skip it")
-		// return nil
-	}
-
-	driver := CSIDriver{
-		Name:    name,
-		Version: v,
-		NodeID:  nodeID,
-	}
-
-	return &driver
-}
-
-func (d *CSIDriver) ValidateControllerServiceRequest(c csi.ControllerServiceCapability_RPC_Type) error {
-	if c == csi.ControllerServiceCapability_RPC_UNKNOWN {
-		return nil
-	}
-
-	for _, cap := range d.Cap {
-		if c == cap.GetRpc().GetType() {
-			return nil
-		}
-	}
-	return status.Error(codes.InvalidArgument, c.String())
-}
-
-func (d *CSIDriver) AddControllerServiceCapabilities(cl []csi.ControllerServiceCapability_RPC_Type) {
-	csc := make([]*csi.ControllerServiceCapability, 0, len(cl))
-
-	for _, c := range cl {
-		klog.Infof("Enabling controller service capability: %v", c.String())
-		csc = append(csc, NewControllerServiceCapability(c))
-	}
-	d.Cap = csc
-}
-
-func (d *CSIDriver) AddNodeServiceCapabilities(nl []csi.NodeServiceCapability_RPC_Type) {
-	nsc := make([]*csi.NodeServiceCapability, 0, len(nl))
-	for _, n := range nl {
-		klog.V(2).Infof("Enabling node service capability: %v", n.String())
-		nsc = append(nsc, NewNodeServiceCapability(n))
-	}
-	d.NSCap = nsc
-}
-
-func (d *CSIDriver) AddVolumeCapabilityAccessModes(vc []csi.VolumeCapability_AccessMode_Mode) {
-	vca := make([]*csi.VolumeCapability_AccessMode, 0, len(vc))
-	for _, c := range vc {
-		klog.Infof("Enabling volume access mode: %v", c.String())
-		vca = append(vca, NewVolumeCapabilityAccessMode(c))
-	}
-	d.VC = vca
-}
-
-func (d *CSIDriver) GetVolumeCapabilityAccessModes() []*csi.VolumeCapability_AccessMode {
-	return d.VC
-}
